+++ app/src/main/java/com/zed/wannawatch/services/api/models/tmdb/MovieDetailResult.kt
@@ -1,4 +1,4 @@
-package com.zed.wannawatch.services.api.models
+package com.zed.wannawatch.services.api.models.tmdb
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.json.JsonObject
