+++ app/src/main/java/com/zed/wannawatch/services/api/models/tmdb/MovieResult.kt
@@ -1,20 +1,20 @@
-package com.zed.wannawatch.services.api.models
+package com.zed.wannawatch.services.api.models.tmdb
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class TvResult(
+data class MovieResult(
     val adult: Boolean,
     val backdrop_path: String?,
-    val first_air_date: String,
     val genre_ids: List<Int>,
     val id: Int,
-    val name: String,
-    val origin_country: List<String>,
     val original_language: String,
-    val original_name: String,
+    val original_title: String,
     val overview: String,
     val popularity: Double,
     val poster_path: String?,
+    val release_date: String,
+    val title: String,
+    val video: Boolean,
     val vote_average: Double,
     val vote_count: Int
 )
