+++ app/src/main/java/com/zed/wannawatch/services/api/models/tmdb/TMDBSearchResult.kt
@@ -1,4 +1,4 @@
-package com.zed.wannawatch.services.api.models
+package com.zed.wannawatch.services.api.models.tmdb
 import kotlinx.serialization.Serializable
 
 @Serializable
